Upload (not working with several yandex api keys)

+++ bot.py
@@ -2,7 +2,7 @@ import telebot
 import config
 import requests
 import vk
-from random import choice
+import random
 from voice import say, recognize
 import os
 import time
@@ -15,21 +15,11 @@ vk_api = vk.API(session)
 
 @bot.message_handler(commands=['say'])
 def say_m(message):
-    themes = ['петрович', 'вовочка', 'мужик', 'петька', 'a', 'девушка', '']
-    records = vk_api.wall.search(domain='baneks', query=choice(themes), owners_only=1, v=5.68, count=100)
-    # for i in range(10):
-    try:
-        print('sending')
-        bot.send_message(message.chat.id, 'sending')
-        random_record = choice(records['items'])
-        file = say(random_record['text'])
-        audio = open(file, 'rb')
-        bot.send_audio(message.chat.id, audio)
-        print('done')
-        path = os.path.join(os.path.abspath(os.path.dirname(__file__)), file)
-        os.remove(path)
-    except:
-        pass
+    file = open('ids.txt', 'r')
+    a = file.readlines()
+    anek = random.choice(a).strip()
+    print(anek)
+    bot.send_audio(message.chat.id, anek)
 
 @bot.message_handler(commands=['help'])
 def help(message):
@@ -41,26 +31,45 @@ def start(message):
 
 @bot.message_handler(commands=['upload'])
 def upload(message):
-    try:
-        n = int(message.text[8:])
-    except:
-        bot.send_message(message.chat.id, 'enter number')
-        return
-    anekdots = aneks.get(n)
-
-    for i in range(len(anekdots)):
+    if message.chat.id == config.admin_chat:
         try:
-            files = open('ids.txt', 'a')
-            file = say(anekdots[i])
-            audio = open(file, 'rb')
-            msg = bot.send_audio(message.chat.id, audio)
-            print(msg.audio.file_id, file=files)
-            files.close()
-            path = os.path.join(os.path.abspath(os.path.dirname(__file__)), file)
-            os.remove(path)
-            time.sleep(1)
+            n = int(message.text[8:])
         except:
-            print('SOME STRANGE ERROR', i)
+            bot.send_message(message.chat.id, 'enter number')
+            return
+        anekdots = aneks.get(n)
+
+        number = 0
+        errors = 0
+        start = 3844
+        for i in range(start, len(anekdots)):
+            try:
+                token = config.token_yandex[number]
+                errors = 0
+                files = open('ids.txt', 'a')
+                file = say(anekdots[i], token)
+                audio = open(file, 'rb')
+                msg = bot.send_audio(message.chat.id, audio)
+                print(msg.audio.file_id, file=files)
+                files.close()
+                path = os.path.join(os.path.abspath(os.path.dirname(__file__)), file)
+                os.remove(path)
+                print(i, 'ready')
+                time.sleep(1)
+            except:
+                print('SOME STRANGE ERROR', i)
+                errors += 1
+                if errors >= 5:
+                    print('OOOps, lots of errors, using the next token and sleeping 20 secs')
+                    if number >= len(config.token_yandex):
+                        print('no more tokens((')
+                        files.close()
+                        exit()
+                    number += 1
+                    time.sleep(20)
+    else:
+        bot.send_message(message.chat.id, 'You-re not admin')
+
 
 if __name__ == '__main__':
     bot.polling(none_stop=True)
